// app_web/client-sn/interfaces/ResultadoScreening.ts
// Generated by https://quicktype.io

import { RespuestaScreening } from "./RespuestaScreening";
import { TipoScreening } from "./TipoScreening";

export interface ResultadoScreening {
    id: number;
    rela_screening: number;
    rela_paciente: number;
    rela_medico: number;
    result_screening: number;
    fecha_alta: string;
    tipo_screening: TipoScreening;
    respuestas?: RespuestaScreening[]
}